python/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.110.1
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
h11==0.14.0
idna==3.6
psycopg-binary==3.1.18
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
